second

+++ yatube/posts/views.py
@@ -57,7 +57,7 @@ def profile(request, username):
         Follow.objects.filter(
             user=request.user,
             author=author,
-        )
+        ).exists()
     )
     context = {
         'author': author,
@@ -177,16 +177,12 @@ def follow_index(request):
 @login_required
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
-    if author != request.user:
+    if author == request.user or Follow.objects.filter(
+            user=request.user, author=author).exists():
 
-        if not Follow.objects.filter(
-            user=request.user,
-            author=author,
-        ).exists():
-            Follow.objects.create(
-                user=request.user,
-                author=author
-            )
+        return redirect('posts:profile', username)
+
+    Follow.objects.create(user=request.user, author=author)
 
     return redirect('posts:profile', username)
 
@@ -194,10 +190,6 @@ def profile_follow(request, username):
 @login_required
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
-    if author != request.user:
-        Follow.objects.filter(
-            user=request.user,
-            author=author,
-        ).delete()
+    Follow.objects.filter(user=request.user, author=author).delete()
 
     return redirect('posts:profile', username)
